setUserData in database service
